feat: TableGroup이 가지고 있는 Order 제거
